added julia package and sysimage precompilation
