+++ CatnipJulia/Project.toml
@@ -1,0 +1,13 @@
+name = "CatnipJulia"
+uuid = "5f57de80-edc4-450b-b0f3-8aa9fe896468"
+authors = ["noriah <vix@noriah.dev> and contributors"]
+version = "1.0.0-DEV"
+
+[deps]
+Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
+DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
+GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
+PackageCompiler = "9b87118b-4619-50d2-8e1e-99f35a4d4d9d"
+
+[compat]
+julia = "1"
