+++ CatnipJulia/precompile_script.jl
@@ -1,0 +1,3 @@
+using CatnipJulia
+
+run_catnip(; timeout=true)
