+++ CatnipJulia/src/CatnipJulia.jl
@@ -1,0 +1,11 @@
+module CatnipJulia
+
+using Colors
+using DataStructures: CircularBuffer
+using GLMakie
+
+include("run_catnip.jl")
+
+export run_catnip
+
+end
